Revert "Ensure that RSSI is not casted to an int (#1164)"
This reverts commit 14b21b5357ce9da2176653ffa25378c35dddf43d.

+++ LoRaEngine/LoraKeysManagerFacade/CheckPreferredGateway/PreferredGatewayTableItem.cs
@@ -7,9 +7,9 @@ namespace LoraKeysManagerFacade
     {
         public string GatewayID { get; private set; }
 
-        public double Rssi { get; private set; }
+        public int Rssi { get; private set; }
 
-        public PreferredGatewayTableItem(string gatewayID, double rssi)
+        public PreferredGatewayTableItem(string gatewayID, int rssi)
         {
             GatewayID = gatewayID;
             Rssi = rssi;
@@ -34,7 +34,7 @@ namespace LoraKeysManagerFacade
             if (values?.Length != 2)
                 return null;
 
-            if (!double.TryParse(values[1], out var rssi))
+            if (!int.TryParse(values[1], out var rssi))
                 return null;
 
             return new PreferredGatewayTableItem(values[0], rssi);
